refactor(RacingCarSerivce): JudgementGenerator 생성
