+++ src/main/java/racingcar/service/RacingCarService.java
@@ -2,34 +2,33 @@ package racingcar.service;
 
 import java.util.List;
 import racingcar.domain.Car;
-import racingcar.domain.CarNameSeperator;
 import racingcar.domain.Cars;
-import racingcar.utils.RandomNumberGenerator;
+import racingcar.utils.CarNameSeperator;
+import racingcar.utils.JudgementGenerator;
+import racingcar.utils.JudgementRandomGenerator;
 
 public class RacingCarService {
 
     private final CarNameSeperator carNameSeperator;
-    private final RandomNumberGenerator randomNumberGenerator;
+    private final JudgementGenerator judgementGenerator;
 
     public RacingCarService() {
         this.carNameSeperator = new CarNameSeperator();
-        this.randomNumberGenerator = new RandomNumberGenerator();
+        this.judgementGenerator = new JudgementRandomGenerator();
     }
 
     public Cars saveCarName(final String input) {
         List<String> seperatedCarNames = carNameSeperator.seperateCarNames(input);
-        Cars cars = new Cars(seperatedCarNames.stream().map(Car::new).toList());
-        return cars;
+        return new Cars(seperatedCarNames.stream().map(Car::new).toList());
     }
 
-    public Cars moveCar(final Cars cars) {
+    public void moveCar(final Cars cars) {
         judgeMoveCar(cars);
-        return cars;
     }
 
     private void judgeMoveCar(final Cars cars) {
         for (Car car : cars.getCars()) {
-            if (randomNumberGenerator.isIncreseDistance()) {
+            if (judgementGenerator.isIncreseDistance()) {
                 car.increaseDistance();
             }
         }
