+++ src/main/java/racingcar/utils/CarNameSeperator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.utils;
 
 import java.util.ArrayList;
 import java.util.List;
