+++ src/main/java/racingcar/utils/JudgementGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.utils;
+
+public interface JudgementGenerator {
+    boolean isIncreseDistance();
+}
