+++ src/main/java/racingcar/utils/JudgementRandomGenerator.java
@@ -2,7 +2,7 @@ package racingcar.utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-public class RandomNumberGenerator {
+public class JudgementRandomGenerator implements JudgementGenerator {
 
     private static final int MIN_RANGE = 0;
     private static final int MAX_RANGE = 9;
